feat(validation): add custom validation and improve entity relationships

+++ src/main/java/fiap/com/br/graus/projection/ProdutoSummary.java
@@ -1,15 +1,16 @@
 package fiap.com.br.graus.projection;
 
+import java.math.BigDecimal;
+
 public interface ProdutoSummary {
 
     Long getId();
 
     String getNome();
 
-    Double getPreco();
+    BigDecimal getPreco();
 
     String getTamanho();
 
     Boolean getAtivo();
-
 }
